requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.41
python-dotenv==1.1.0
openai==1.86.0
requests==2.32.4
gunicorn==21.2.0
psycopg2-binary==2.9.9
Flask-Migrate==4.0.5
tqdm==4.67.1
certifi==2025.6.15
charset-normalizer==3.4.2
idna==3.10
urllib3==2.4.0
Werkzeug==3.1.3
Jinja2==3.1.6
itsdangerous==2.2.0
MarkupSafe==3.0.2
click==8.2.1
blinker==1.9.0
